requirements.txt
google-api-python-client==2.56.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
pytest==7.1.2
selenium==4.3.0
